// packages/cactus-plugin-ledger-connector-fabric/src/main/typescript/generated/protos/common/ack_pb.ts
// Copyright IBM Corp. All Rights Reserved.
//
// SPDX-License-Identifier: Apache-2.0

// @generated by protoc-gen-es v2.2.2 with parameter "target=ts"
// @generated from file common/ack.proto (package common.ack, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file common/ack.proto.
 */
export const file_common_ack: GenFile = /*@__PURE__*/
  fileDesc("ChBjb21tb24vYWNrLnByb3RvEgpjb21tb24uYWNrIm8KA0FjaxImCgZzdGF0dXMYAiABKA4yFi5jb21tb24uYWNrLkFjay5TVEFUVVMSEgoKcmVxdWVzdF9pZBgDIAEoCRIPCgdtZXNzYWdlGAQgASgJIhsKBlNUQVRVUxIGCgJPSxAAEgkKBUVSUk9SEAFCdgoub3JnLmh5cGVybGVkZ2VyLmNhY3RpLndlYXZlci5wcm90b3MuY29tbW9uLmFja1pEZ2l0aHViLmNvbS9oeXBlcmxlZGdlci1jYWN0aS9jYWN0aS93ZWF2ZXIvY29tbW9uL3Byb3Rvcy1nby92Mi9jb21tb25iBnByb3RvMw");

/**
 * This message respresents "ACKs" sent between relay-relay,
 * relay-driver and relay-network
 *
 * @generated from message common.ack.Ack
 */
export type Ack = Message<"common.ack.Ack"> & {
  /**
   * @generated from field: common.ack.Ack.STATUS status = 2;
   */
  status: Ack_STATUS;

  /**
   * @generated from field: string request_id = 3;
   */
  requestId: string;

  /**
   * an error can have an associated string
   * this is the best way to represent this in protobuf
   *
   * @generated from field: string message = 4;
   */
  message: string;
};

/**
 * Describes the message common.ack.Ack.
 * Use `create(AckSchema)` to create a new message.
 */
export const AckSchema: GenMessage<Ack> = /*@__PURE__*/
  messageDesc(file_common_ack, 0);

/**
 * @generated from enum common.ack.Ack.STATUS
 */
export enum Ack_STATUS {
  /**
   * @generated from enum value: OK = 0;
   */
  OK = 0,

  /**
   * @generated from enum value: ERROR = 1;
   */
  ERROR = 1,
}

/**
 * Describes the enum common.ack.Ack.STATUS.
 */
export const Ack_STATUSSchema: GenEnum<Ack_STATUS> = /*@__PURE__*/
  enumDesc(file_common_ack, 0, 0);